Refactoring directory structures
